requirements.txt
streamlit==1.32.0
pydantic==2.6.0
sentence-transformers==2.5.0
faiss-cpu==1.8.0
PyMuPDF==1.26.6
openai==1.14.0
httpx==0.27.0
scikit-learn==1.4.0
numpy==1.26.0
